requirements.txt
# requirements.txt
alpaca-py==0.11.0
numpy==1.24.3
pandas==2.0.3
torch==2.1.0
scikit-learn==1.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
textblob==0.17.1
requests==2.31.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0
python-json-logger==2.0.7
redis==5.0.0
prometheus-client==0.17.1
boto3==1.28.3